@@ -8,10 +8,14 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "ethers": "^5.6.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1"
+    "@tanstack/react-query": "^5.74.11",
+    "ethers": "^5.8.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-icons": "^4.3.1",
+    "react-router-dom": "^6.3.0",
+    "viem": "^2.28.1",
+    "wagmi": "^2.15.1"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^1.0.7",
