@@ -1,6 +1,6 @@
 const Footer = () => {
     return(
-        <h1>Footer</h1>
+        <div>Created by me</div>
     )
 }
 
